+++ tests/algorithms/test_linalg.py
@@ -3,7 +3,11 @@ import pytest
 
 from qubrabench.algorithms.linalg import qlsa_query_count, solve
 from qubrabench.benchmark import track_queries
-from qubrabench.datastructures.matrix import Qndarray, block_encode_matrix
+from qubrabench.datastructures.qndarray import (
+    Qndarray,
+    block_encode_matrix,
+    state_preparation_unitary,
+)
 
 
 def random_instance(rng, N: int) -> tuple[Qndarray, Qndarray]:
@@ -15,9 +19,11 @@ def random_instance(rng, N: int) -> tuple[Qndarray, Qndarray]:
 @pytest.mark.parametrize("N", [5, 10, 15, 20, 25, 30])
 def test_solve(rng, N: int):
     A, b = random_instance(rng, N)
+
     enc_A = block_encode_matrix(A, eps=0)
-    enc_b = block_encode_matrix(b, eps=0)
+    enc_b = state_preparation_unitary(b, eps=0)
     enc_y = solve(enc_A, enc_b)
+
     np.testing.assert_allclose(enc_A.matrix @ enc_y.matrix, enc_b.matrix)
 
 
@@ -25,7 +31,7 @@ def test_solve_stats(rng):
     N = 10
     A, b = random_instance(rng, N)
     enc_A = block_encode_matrix(A, eps=0)
-    enc_b = block_encode_matrix(b, eps=0)
+    enc_b = state_preparation_unitary(b, eps=0)
     enc_y = solve(enc_A, enc_b, error=1e-5)
 
     with track_queries() as tracker:
