+++ tests/datastructures/test_matrix.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from qubrabench.benchmark import QueryStats, track_queries
-from qubrabench.datastructures.matrix import Qndarray
+from qubrabench.datastructures.qndarray import Qndarray
 
 
 def test_qmatrix_iterate(rng):
